Add person detection example
This patch includes the source code which utlizes the 1 outer
camera to capture 160*120 RGB565 format image and then preprocessed
to a 96*96 grayscale image for feeding to the model. The
preprocessed images are shown on top screen for a better
visualization. The bottom screen shows the inferenced result.

One screenshot and a footage are included in the patch, together
with the updated README.md.